Add explicit type cast to avoid gcc release build warning

+++ src/civetweb.c
@@ -2409,7 +2409,8 @@ skip_quoted(char **buf,
 	if (*end_word == '\0') {
 		*buf = end_word;
 	} else {
-		end_whitespace = end_word + 1 + strspn(end_word + 1, whitespace);
+		end_whitespace =
+		    end_word + 1u + (unsigned int)strspn(end_word + 1, whitespace);
 
 		for (p = end_word; p < end_whitespace; p++) {
 			*p = '\0';
